requirements.txt
matplotlib==2.1.0
numpy==1.13.3
Pillow==4.2.1
Shapely==1.6.2
scipy
imutils
cv2